+++ wemake_python_styleguide/violations/refactoring.py
@@ -1031,7 +1031,7 @@ class InCompareWithSingleItemContainerViolation(ASTViolation):
 
     """
 
-    error_template = 'Found wrong "in" compare with single item container'
+    error_template = 'Found wrong `in` compare with single item container'
     code = 525